Generate manpage with sphinx and add it to package
This commit addresses issue #162 which suggested to add a manpage to the tldr
python client. The changes in the `tldr.py` file where necessary, because
sphinx needs an easy way to obtain the `Argparser`-object. I also added a
Makefile, so the manpage can easily recreated.

You need the sphinx-build tool and the argparse extension. Both can
usually installed via. your package-manager. In my case the packages
where called `python-sphinx` and `python-sphinx-argparse`.

Closes #162