Search only through local cache
This way, we save a lot of time and prevent the program from being
locked up downloading lots of pages and searching through them.
However, the final page may still be downloaded in case cache isn't up
to date enough. This should be fixed by adding a check to see if cache
is up to date or not, but the provided function didn't make a whole
bunch of sense to me. Hopefully someone else can provide needed info
here.